Made it to where if you are not appart of the Team you can not post a discussion to that Team
